+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog of threedigrid-builder
 
 - Bugfix: Small fix adding groundwater cells.
 
+- Bugfix: also need to evaluate embedded nodes for connection node mapping for zero-d surface maps.
+
 
 0.9.0 (2021-12-15)
 ------------------